+++ src/main/java/com/kshrd/autopilot/service/DeploymentAppService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface DeploymentAppService {
     DeploymentAppDto createDeploymentApp(DeploymentAppRequest request) throws IOException, InterruptedException;
-    List<DeploymentAppDto>getAllDeploymentApps(Integer project_id);
+    List<DeploymentAppDto>getAllDeploymentApps(Long project_id);
 }
